Renaming some files for naming conventions
