+++ src/graph/KokkosGraph_graph_color.hpp
@@ -105,7 +105,7 @@ void graph_color_symbolic(
  
   case COLORING_SPGEMM:
   case COLORING_D2_MATRIX_SQUARED:
-    //std::cout << ">>> WCMCLEN graph_color_symbolic (KokkosGraph_GraphColor.hpp) [ COLORING_SPGEMM / COLORING_D2_MATRIX_SQUARED ]" << std::endl;
+    //std::cout << ">>> WCMCLEN graph_color_symbolic (KokkosGraph_graph_color.hpp) [ COLORING_SPGEMM / COLORING_D2_MATRIX_SQUARED ]" << std::endl;
 
     if (handle->get_handle_exec_space() == KokkosKernels::Impl::Exec_CUDA) {
         typedef typename Impl::GraphColor_EB <typename KernelHandle::GraphColoringHandleType, lno_row_view_t_, lno_nnz_view_t_> EBGraphColoringSPGEMM;
@@ -119,7 +119,7 @@ void graph_color_symbolic(
 //  case COLORING_D2_WCMCLEN:
 //    {
 //      std::ostringstream os;
-//      os << ">>> WCMCLEN graph_color_symbolic (KokkosGraph_GraphColor.hpp) COLORING_D2_WCMCLEN not implemented";
+//      os << ">>> WCMCLEN graph_color_symbolic (KokkosGraph_graph_color.hpp) COLORING_D2_WCMCLEN not implemented";
 //      Kokkos::Impl::throw_runtime_exception(os.str());
 //    }
 //    break;
@@ -142,6 +142,16 @@ void graph_color_symbolic(
   gch->set_vertex_colors(colors_out);
 }
 
+template <class KernelHandle,typename lno_row_view_t_, typename lno_nnz_view_t_>
+void graph_color(
+    KernelHandle *handle,
+    typename KernelHandle::nnz_lno_t num_rows,
+    typename KernelHandle::nnz_lno_t num_cols,
+    lno_row_view_t_ row_map,
+    lno_nnz_view_t_ entries,
+    bool is_symmetric = true){
+  graph_color_symbolic(handle,num_rows,num_cols,row_map,entries,is_symmetric);
+}
 
 
 template <class KernelHandle,typename lno_row_view_t_, typename lno_nnz_view_t_>
@@ -153,7 +163,7 @@ void graph_color(
     lno_nnz_view_t_ entries,
     bool is_symmetric = true)
 {
-  //std::cout << ">>> WCMCLEN graph_color (KokkosGraph_GraphColor.hpp)" << std::endl;
+  //std::cout << ">>> WCMCLEN graph_color (KokkosGraph_graph_color.hpp)" << std::endl;
   graph_color_symbolic(handle, num_rows, num_cols, row_map, entries, is_symmetric);
 }
