+++ src/sparse/impl/KokkosSparse_gauss_seidel_impl.hpp
@@ -47,7 +47,7 @@
 #include <impl/Kokkos_Timer.hpp>
 #include <Kokkos_Sort.hpp>
 #include <Kokkos_MemoryTraits.hpp>
-#include "KokkosGraph_GraphColor.hpp"
+#include "KokkosGraph_graph_color.hpp"
 
 #ifndef _KOKKOSGSIMP_HPP
 #define _KOKKOSGSIMP_HPP
